fix: move ck files to lib
