+++ frappe/public/js/lib/ckeditor.js
