+++ frappe/public/less/variables.less
@@ -1,4 +1,4 @@
-@brand-primary:			#5e64ff;
+@brand-primary:			#80bd9e;
 @erpnext-blue:			#7575ff;
 @border-color:			#d1d8dd;
 @light-border-color:		#ebeff2;
